Include native price for LP tokens parsing
