--- src/components/expanded-state/liquidity-pool/index.js
@@ -1,1 +1,0 @@
-export { default as LiquidityPoolExpandedStateHeader } from './LiquidityPoolExpandedStateHeader';
