+++ src/hoc/assetSelectors.js
@@ -1,13 +1,7 @@
 import { filter, get, groupBy, isEmpty, isNil, map, toNumber } from 'lodash';
 import { createSelector } from 'reselect';
 import { sortList } from '../helpers/sortList';
-import {
-  add,
-  convertAmountAndPriceToNativeDisplay,
-  convertAmountToNativeDisplay,
-  convertAmountToPercentageDisplay,
-} from '../helpers/utilities';
-import { isLowerCaseMatch } from '../utils';
+import { parseAssetsNativeWithTotals } from '../parsers/accounts';
 
 const EMPTY_ARRAY = [];
 
@@ -53,7 +47,10 @@ const sortAssetsByNativeAmount = (
   let assetsNativePrices = updatedAssets;
   let total = null;
   if (!isEmpty(assetsNativePrices)) {
-    const parsedAssets = parseAssetsNative(assetsNativePrices, nativeCurrency);
+    const parsedAssets = parseAssetsNativeWithTotals(
+      assetsNativePrices,
+      nativeCurrency
+    );
     assetsNativePrices = parsedAssets.assetsNativePrices;
     total = parsedAssets.total;
   }
@@ -94,50 +91,6 @@ const sortAssetsByNativeAmount = (
 
 const groupAssetsByMarketValue = assets =>
   groupBy(assets, ({ native }) => (isNil(native) ? 'noValue' : 'hasValue'));
-
-const parseAssetsNative = (assets, nativeCurrency) => {
-  let assetsNative = assets;
-  assetsNative = map(assets, asset => {
-    const assetNativePrice = get(asset, 'price');
-    if (isNil(assetNativePrice)) {
-      return asset;
-    }
-
-    const priceUnit = get(assetNativePrice, 'value', 0);
-    const nativeDisplay = convertAmountAndPriceToNativeDisplay(
-      get(asset, 'balance.amount', 0),
-      priceUnit,
-      nativeCurrency
-    );
-    return {
-      ...asset,
-      native: {
-        balance: nativeDisplay,
-        change: isLowerCaseMatch(get(asset, 'symbol'), nativeCurrency)
-          ? null
-          : assetNativePrice.relative_change_24h
-          ? convertAmountToPercentageDisplay(
-              assetNativePrice.relative_change_24h
-            )
-          : '',
-        price: {
-          amount: priceUnit,
-          display: convertAmountToNativeDisplay(priceUnit, nativeCurrency),
-        },
-      },
-    };
-  });
-  const totalAmount = assetsNative.reduce(
-    (total, asset) => add(total, get(asset, 'native.balance.amount', 0)),
-    0
-  );
-  const totalDisplay = convertAmountToNativeDisplay(
-    totalAmount,
-    nativeCurrency
-  );
-  const total = { amount: totalAmount, display: totalDisplay };
-  return { assetsNativePrices: assetsNative, total };
-};
 
 export const sortAssetsByNativeAmountSelector = createSelector(
   [
